+++ src/main/kotlin/net/ccbluex/liquidbounce/features/module/modules/exploit/servercrasher/ModuleServerCrasher.kt
@@ -43,7 +43,7 @@ object ModuleServerCrasher : ClientModule("ServerCrasher", Category.EXPLOIT, dis
         super.enable()
     }
 
-    val exploitChoices = choices<Choice>("Exploit", Log4jChatExploit, arrayOf(
+    val exploitChoices = choices<Choice>("Exploit", Bundle1214Exploit, arrayOf(
         // Version Specific
         Bundle1214Exploit,
         PaperCompletion1204Exploit,
