+++ src/main/kotlin/net/ccbluex/liquidbounce/utils/item/ItemExtensions.kt
@@ -105,6 +105,29 @@ val ItemStack.isFood: Boolean
 val ItemStack.foodComponent: FoodComponent?
     get() = this.get(DataComponentTypes.FOOD)
 
+private val BUNDLE_ITEMS = setOf(
+    Items.BUNDLE,
+    Items.WHITE_BUNDLE,
+    Items.ORANGE_BUNDLE,
+    Items.MAGENTA_BUNDLE,
+    Items.LIGHT_BLUE_BUNDLE,
+    Items.YELLOW_BUNDLE,
+    Items.LIME_BUNDLE,
+    Items.PINK_BUNDLE,
+    Items.GRAY_BUNDLE,
+    Items.LIGHT_GRAY_BUNDLE,
+    Items.CYAN_BUNDLE,
+    Items.PURPLE_BUNDLE,
+    Items.BLUE_BUNDLE,
+    Items.BROWN_BUNDLE,
+    Items.GREEN_BUNDLE,
+    Items.RED_BUNDLE,
+    Items.BLACK_BUNDLE
+)
+
+val ItemStack.isBundle
+    get() = this.item in BUNDLE_ITEMS
+
 fun isHotbarSlot(slot: Int) = slot == 45 || slot in 36..44
 
 val MiningToolItem.type: Int
